Backend/pyproject.toml
[tool.poetry]
name = "mtaawetu-backend"
version = "0.1.0"
description = "backend for the mtaa wetu community engagement app / data store"
authors = ["John Ngugi <johnngugi0407@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


[tool.poetry.group.dev.dependencies]
django = "^5.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.5.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
